extract dataclasses for agri2018
